Final/requirements.txt
azure_storage==0.37.0
docxtpl==0.16.6
Flask==2.3.2
mac_vendor_lookup==0.1.12
matplotlib==3.7.1
netifaces==0.11.0
networkx==3.1
Pillow==9.5.0
psutil==5.9.5
pymongo==4.3.3
scapy==2.5.0
python-nmap==0.7.1
azure==4.0.0